moved the display of transit trip to a new component (DisplayTransitWalk)for a better overview
